requirements.txt
rerun-sdk==0.20.1
pyzmq==27.0.0
logging_mp==0.1.5
onnxruntime==1.22.1
onnx
pynput==1.8.1

# Isaac-GR00T集成依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
tqdm>=4.64.0
pyyaml>=6.0
onnxruntime-gpu>=1.22.1  # GPU加速（可选）